Fixed Zoom Scaling ChartComposite.java
When Chart is scaled, zooming did not work correctly when useBuffer is false. But when useBuffer is true, zoomRectangle was not drawn on the buffer, one can see that when setting background to black and Outline color to white. These changes fix both issues and also made the zoomRectangle 2 pixels to avoid flickering when useBuffer is true